🔧 refactor(__main__.py, main.py): move setup_static_files function to main.py and rename create_app to setup_app ✨ feat(__main__.py, main.py): add support for a custom static files directory to be passed as an argument to the app The `setup_static_files` function has been moved from `__main__.py` to `main.py` to improve code organization. The function has also been renamed to `setup_app` to better reflect its purpose. The `create_app` function has been renamed to `setup_app` to follow the naming convention of the new function. The `setup_app` function now accepts an optional argument `static_files_dir` which allows the user to specify a custom directory for static files. This improves the flexibility of the app as it can now be run with a custom frontend.
